+++ Modules/Intent.Modules.VisualStudio.Projects/Api/SolutionFolderModel.cs
@@ -123,6 +123,11 @@ namespace Intent.Modules.VisualStudio.Projects.Api
             .Select(x => new CSharpProjectNETModel(x))
             .ToList();
 
+        public IList<JavaScriptProjectModel> JavaScriptProjects => _element.ChildElements
+            .GetElementsOfType(JavaScriptProjectModel.SpecializationTypeId)
+            .Select(x => new JavaScriptProjectModel(x))
+            .ToList();
+
         public IList<ConsoleAppNETCoreModel> ConsoleAppNETCores => _element.ChildElements
             .GetElementsOfType(ConsoleAppNETCoreModel.SpecializationTypeId)
             .Select(x => new ConsoleAppNETCoreModel(x))