+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/VisualStudio2015Solution/Registrations.cs
@@ -22,7 +22,13 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.VisualStudio2015Solutio
             var vsSolutions = _metadataManager.VisualStudio(application).GetVisualStudioSolutionModels();
             foreach (var vsSolution in vsSolutions)
             {
-                var projects = _metadataManager.GetAllProjectModels(application).Where(x => x.Solution.Id == vsSolution.Id).ToList();
+                var projects = _metadataManager
+                    .GetAllProjectModels(application)
+                    .Cast<IVisualStudioSolutionProject>()
+                    .Concat(_metadataManager.VisualStudio(application).GetJavaScriptProjectModels())
+                    .Where(x => x.Solution.Id == vsSolution.Id)
+                    .ToList();
+
                 registry.RegisterApplicationTemplate(VisualStudio2015SolutionTemplate.Identifier, () => new VisualStudio2015SolutionTemplate(application, vsSolution, projects));
             }
         }