+++ Modules/Intent.Modules.VisualStudio.Projects/release-notes.md
@@ -4,7 +4,9 @@
 - Improvement: Deprecated ".NET Core" project types are no longer visible on the context menu, use `C# Project (.NET)` instead and change the `SDK` and other options on it to create equivalent projects.
 - Improvement: Options to create ".NET Framework" projects are now hidden by default, to have these options show use toggle the `Enable .NET Framework project creation` application setting.
 - Improvement: `Add .NET Core Version` and `Add .NET Framework Version` options are now hidden unless you have the `Intent.ModuleBuilder` module installed.
+- Improvement: Solution Folders will now be generated in `.sln` files for templates with a relative output location when their output is in the package root or a Solution Folder.
 - Fixed: NuGet package removals weren't being processed.
+- Fixed: Certain `.sln` files could cause exceptions to be thrown during Software Factory execution.
 
 ### Version 3.5.1